// model/hangarType.ts
export type Hangar = {
  id: number;
  liveryId: number;
  cost: number;
  location: string;
  statusEngine: number;
  statusHull: number;
  airframe: number;
  currentFuel: number;
  ownerId: number;
  status: number;
  insurance: number;
  tailNumber: string;
  source: number;
  lastPayDate: number;
};